docgpt_main-main/docgpt_main-main/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv
python-multipart==0.0.6
numpy
tiktoken==0.5.2
faiss-cpu==1.7.4
langchain==0.0.350
openai
pypdf